+++ internal/execution/executor.go
@@ -1,12 +1,8 @@
 package execution
 
 import (
-	"bytes"
 	"context"
-	"encoding/json"
 	"fmt"
-	"os"
-	"os/exec"
 	"strconv"
 	"time"
 
@@ -24,20 +20,6 @@ import (
 
 // maxSliceJsonBytes is the max sum of a resource slice's manifests.
 const maxSliceJsonBytes = 1024 * 512
-
-type Env struct {
-	CompositionName      string
-	CompositionNamespace string
-	SynthesisUUID        string
-}
-
-func LoadEnv() *Env {
-	return &Env{
-		CompositionName:      os.Getenv("COMPOSITION_NAME"),
-		CompositionNamespace: os.Getenv("COMPOSITION_NAMESPACE"),
-		SynthesisUUID:        os.Getenv("SYNTHESIS_UUID"),
-	}
-}
 
 type Executor struct {
 	Reader  client.Reader
@@ -232,47 +214,4 @@ func (e *Executor) updateComposition(ctx context.Context, oldComp *apiv1.Composi
 		logger.V(0).Info("composition status has been updated following successful synthesis")
 		return nil
 	})
-}
-
-type SynthesizerHandle func(context.Context, *apiv1.Synthesizer, *krmv1.ResourceList) (*krmv1.ResourceList, error)
-
-func NewExecHandler() SynthesizerHandle {
-	return func(ctx context.Context, s *apiv1.Synthesizer, rl *krmv1.ResourceList) (*krmv1.ResourceList, error) {
-		stdin := &bytes.Buffer{}
-		stdout := &bytes.Buffer{}
-
-		err := json.NewEncoder(stdin).Encode(rl)
-		if err != nil {
-			return nil, err
-		}
-
-		command := s.Spec.Command
-		if len(command) == 0 {
-			command = []string{"synthesize"}
-		}
-
-		if s.Spec.ExecTimeout != nil {
-			var cancel context.CancelFunc
-			ctx, cancel = context.WithTimeout(ctx, s.Spec.ExecTimeout.Duration)
-			defer cancel()
-		}
-
-		cmd := exec.CommandContext(ctx, command[0], command[1:]...)
-		cmd.Env = []string{} // no env
-		cmd.Stdin = stdin
-		cmd.Stderr = os.Stdout // logger uses stderr, so use stdout to avoid race condition
-		cmd.Stdout = stdout
-		err = cmd.Run()
-		if err != nil {
-			return nil, err
-		}
-
-		output := &krmv1.ResourceList{}
-		err = json.NewDecoder(stdout).Decode(output)
-		if err != nil {
-			return nil, err
-		}
-
-		return output, nil
-	}
 }
